Updated Anti-Keypress
I've gone ahead and updated the Anti-Keypress stability. I have made it easier for me to manage and add certain banned keybinds, as ALT+F4, and CTRL+SHIFT+ESC.
Instead of returning back the original output for keypress, I have made it error to further stop these type of scripts that can tamper with your keyboard.
It used to go ahead and return the same response, but make it NOT break. However, this update completely STOPS that.